+++ bwmidimapper/__init__.py
@@ -1,1 +1,3 @@
+"""bwMIDIMapper - A tool to convert MIDI files between different drum mappings."""
+
 from .main import main