+++ CHANGELOG.md
@@ -8,6 +8,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+## [1.2.0] - 2026-02-07
+- Changed maximum line length to 100 (based on PEP 8).
+- Refactored to use AppConfig and MIDIHandler classes.
+
 ## [1.1.0] - 2026-02-04
 - Remove duplicate MIDI notes.